+++ opensearch/elasticsearch-exporter/versions.py
@@ -1,8 +1,0 @@
-versions = [
-    {
-        "product": "1.9.0",
-        "golang": "1.23.9",
-        "stackable-devel": "1.0.0",
-        "promu": "0.17.0",
-    },
-]