+++ opensearch/versions.py
@@ -4,6 +4,6 @@ versions = [
         "java-devel": "21",
         "jdk-base": "21",
         "opensearch/security-plugin": "3.1.0.0",
-        "opensearch/elasticsearch-exporter": "1.9.0",
+        "opensearch/opensearch-prometheus-exporter": "3.1.0.0",
     },
 ]